remove: #44 calculation details as html (#46)
Very crudely remove the calculation details documentation as per #44 


![image](https://github.com/user-attachments/assets/c24c762c-318b-48fd-837c-bbb8de31b620)

As a leftover there's a relatively harmless useless functionality where
clicking the link from e.g. food calculation appends `#food` to the url,
which I'm willing to leave in.

As we don't currently have the resources to update the html view of the
calculation details to the newest version, it's clearer to only provide
the pdf documentation.

Removes the display functionality and translations to reduce the bundle
size of the application, although we might reintroduce these back at a
later time if the need arises.